+++ decompiler/decompile.ts
@@ -1,544 +1,368 @@
-// Decompile MicroPython - a micropython .mpy file decompiler
-// 'M',
-// MPY_VERSION,
-// MPY_FEATURE_ENCODE_FLAGS(MPY_FEATURE_FLAGS_DYNAMIC),
+// <module>: ; test.py
+// 	MP_BC_LOAD_FAST_N 38, 36, 0
+// 	true ; $push(True)
+// 	st.name fancy ; fancy = $pop(0);
+// 	loadname fancy ; $push(fancy);
+// 	bfalse .L36 ; if $pop(0) {
+// 	true ; $push(True)
+// 	st.name x ; x = $pop(0)
+// 	loadname print ; $push(print)
+// 	str asdf ; $push("asdf")
+// 	loadname x ; $push(x)
+// 	call 2 ; $push($pop(2)($pop(0), $pop(1)))
+// .L36: ; }
+// 	pop ; $afterbrpop()
+// 	none ; $push(None)
 
-import { execSync } from "child_process";
-import { existsSync, writeFileSync } from "fs";
+import { readFileSync } from 'fs'
+type fmt = string[][]
 
-const MP_BC_BASE_QSTR_O = 0x10; // LLLLLLSSSDDII---
-const MP_BC_BASE_VINT_E = 0x20; // MMLLLLSSDDBBBBBB
-const MP_BC_BASE_VINT_O = 0x30; // UUMMCCCC--------
-const MP_BC_BASE_JUMP_E = 0x40; // J-JJJJJEEEEF----
-const MP_BC_BASE_BYTE_O = 0x50; // LLLLSSDTTTTTEEFF
-const MP_BC_BASE_BYTE_E = 0x60; // --BREEEYYI------
-let importmap = {
-	MP_BC_LOAD_CONST_FALSE: MP_BC_BASE_BYTE_O + 0x00,
-	MP_BC_LOAD_CONST_NONE: MP_BC_BASE_BYTE_O + 0x01,
-	MP_BC_LOAD_CONST_TRUE: MP_BC_BASE_BYTE_O + 0x02,
-	MP_BC_LOAD_CONST_SMALL_INT: MP_BC_BASE_VINT_E + 0x02, // signed var-int
-	MP_BC_LOAD_CONST_STRING: MP_BC_BASE_QSTR_O + 0x00, // qstr
-	MP_BC_LOAD_CONST_OBJ: MP_BC_BASE_VINT_E + 0x03, // ptr
-	MP_BC_LOAD_NULL: MP_BC_BASE_BYTE_O + 0x03,
-	MP_BC_LOAD_FAST_N: MP_BC_BASE_VINT_E + 0x04, // uint
-	MP_BC_LOAD_DEREF: MP_BC_BASE_VINT_E + 0x05, // uint
-	MP_BC_LOAD_NAME: MP_BC_BASE_QSTR_O + 0x01, // qstr
-	MP_BC_LOAD_GLOBAL: MP_BC_BASE_QSTR_O + 0x02, // qstr
-	MP_BC_LOAD_ATTR: MP_BC_BASE_QSTR_O + 0x03, // qstr
-	MP_BC_LOAD_METHOD: MP_BC_BASE_QSTR_O + 0x04, // qstr
-	MP_BC_LOAD_SUPER_METHOD: MP_BC_BASE_QSTR_O + 0x05, // qstr
-	MP_BC_LOAD_BUILD_CLASS: MP_BC_BASE_BYTE_O + 0x04,
-	MP_BC_LOAD_SUBSCR: MP_BC_BASE_BYTE_O + 0x05,
-	MP_BC_STORE_FAST_N: MP_BC_BASE_VINT_E + 0x06, // uint
-	MP_BC_STORE_DEREF: MP_BC_BASE_VINT_E + 0x07, // uint
-	MP_BC_STORE_NAME: MP_BC_BASE_QSTR_O + 0x06, // qstr
-	MP_BC_STORE_GLOBAL: MP_BC_BASE_QSTR_O + 0x07, // qstr
-	MP_BC_STORE_ATTR: MP_BC_BASE_QSTR_O + 0x08, // qstr
-	MP_BC_STORE_SUBSCR: MP_BC_BASE_BYTE_O + 0x06,
-	MP_BC_DELETE_FAST: MP_BC_BASE_VINT_E + 0x08, // uint
-	MP_BC_DELETE_DEREF: MP_BC_BASE_VINT_E + 0x09, // uint
-	MP_BC_DELETE_NAME: MP_BC_BASE_QSTR_O + 0x09, // qstr
-	MP_BC_DELETE_GLOBAL: MP_BC_BASE_QSTR_O + 0x0a, // qstr
-	MP_BC_DUP_TOP: MP_BC_BASE_BYTE_O + 0x07,
-	MP_BC_DUP_TOP_TWO: MP_BC_BASE_BYTE_O + 0x08,
-	MP_BC_POP_TOP: MP_BC_BASE_BYTE_O + 0x09,
-	MP_BC_ROT_TWO: MP_BC_BASE_BYTE_O + 0x0a,
-	MP_BC_ROT_THREE: MP_BC_BASE_BYTE_O + 0x0b,
-	MP_BC_JUMP: MP_BC_BASE_JUMP_E + 0x02, // rel byte code offset, 16-bit signed, in excess
-	MP_BC_POP_JUMP_IF_TRUE: MP_BC_BASE_JUMP_E + 0x03, // rel byte code offset, 16-bit signed, in excess
-	MP_BC_POP_JUMP_IF_FALSE: MP_BC_BASE_JUMP_E + 0x04, // rel byte code offset, 16-bit signed, in excess
-	MP_BC_JUMP_IF_TRUE_OR_POP: MP_BC_BASE_JUMP_E + 0x05, // rel byte code offset, 16-bit signed, in excess
-	MP_BC_JUMP_IF_FALSE_OR_POP: MP_BC_BASE_JUMP_E + 0x06, // rel byte code offset, 16-bit signed, in excess
-	MP_BC_UNWIND_JUMP: MP_BC_BASE_JUMP_E + 0x00, // rel byte code offset, 16-bit signed, in excess; then a byte
-	MP_BC_SETUP_WITH: MP_BC_BASE_JUMP_E + 0x07, // rel byte code offset, 16-bit unsigned
-	MP_BC_SETUP_EXCEPT: MP_BC_BASE_JUMP_E + 0x08, // rel byte code offset, 16-bit unsigned
-	MP_BC_SETUP_FINALLY: MP_BC_BASE_JUMP_E + 0x09, // rel byte code offset, 16-bit unsigned
-	MP_BC_POP_EXCEPT_JUMP: MP_BC_BASE_JUMP_E + 0x0a, // rel byte code offset, 16-bit unsigned
-	MP_BC_FOR_ITER: MP_BC_BASE_JUMP_E + 0x0b, // rel byte code offset, 16-bit unsigned
-	MP_BC_WITH_CLEANUP: MP_BC_BASE_BYTE_O + 0x0c,
-	MP_BC_END_FINALLY: MP_BC_BASE_BYTE_O + 0x0d,
-	MP_BC_GET_ITER: MP_BC_BASE_BYTE_O + 0x0e,
-	MP_BC_GET_ITER_STACK: MP_BC_BASE_BYTE_O + 0x0f,
-	MP_BC_BUILD_TUPLE: MP_BC_BASE_VINT_E + 0x0a, // uint
-	MP_BC_BUILD_LIST: MP_BC_BASE_VINT_E + 0x0b, // uint
-	MP_BC_BUILD_MAP: MP_BC_BASE_VINT_E + 0x0c, // uint
-	MP_BC_STORE_MAP: MP_BC_BASE_BYTE_E + 0x02,
-	MP_BC_BUILD_SET: MP_BC_BASE_VINT_E + 0x0d, // uint
-	MP_BC_BUILD_SLICE: MP_BC_BASE_VINT_E + 0x0e, // uint
-	MP_BC_STORE_COMP: MP_BC_BASE_VINT_E + 0x0f, // uint
-	MP_BC_UNPACK_SEQUENCE: MP_BC_BASE_VINT_O + 0x00, // uint
-	MP_BC_UNPACK_EX: MP_BC_BASE_VINT_O + 0x01, // uint
-	MP_BC_RETURN_VALUE: MP_BC_BASE_BYTE_E + 0x03,
-	MP_BC_RAISE_LAST: MP_BC_BASE_BYTE_E + 0x04,
-	MP_BC_RAISE_OBJ: MP_BC_BASE_BYTE_E + 0x05,
-	MP_BC_RAISE_FROM: MP_BC_BASE_BYTE_E + 0x06,
-	MP_BC_YIELD_VALUE: MP_BC_BASE_BYTE_E + 0x07,
-	MP_BC_YIELD_FROM: MP_BC_BASE_BYTE_E + 0x08,
-	MP_BC_MAKE_FUNCTION: MP_BC_BASE_VINT_O + 0x02, // uint
-	MP_BC_MAKE_FUNCTION_DEFARGS: MP_BC_BASE_VINT_O + 0x03, // uint
-	MP_BC_MAKE_CLOSURE: MP_BC_BASE_VINT_E + 0x00, // uint; extra byte
-	MP_BC_MAKE_CLOSURE_DEFARGS: MP_BC_BASE_VINT_E + 0x01, // uint; extra byte
-	MP_BC_CALL_FUNCTION: MP_BC_BASE_VINT_O + 0x04, // uint
-	MP_BC_CALL_FUNCTION_VAR_KW: MP_BC_BASE_VINT_O + 0x05, // uint
-	MP_BC_CALL_METHOD: MP_BC_BASE_VINT_O + 0x06, // uint
-	MP_BC_CALL_METHOD_VAR_KW: MP_BC_BASE_VINT_O + 0x07, // uint
-	MP_BC_IMPORT_NAME: MP_BC_BASE_QSTR_O + 0x0b, // qstr
-	MP_BC_IMPORT_FROM: MP_BC_BASE_QSTR_O + 0x0c, // qstr
-	MP_BC_IMPORT_STAR: MP_BC_BASE_BYTE_E + 0x09,
-
-	UNOP_POSITIVE: 0xd0,
-	UNOP_NEGATIVE: 0xd1,
-	UNOP_INVERT: 0xd2,
-	UNOP_NOT: 0xd3,
-
-	BINOP_LESS: 0xd7,
-	BINOP_MORE: 0xd8,
-	BINOP_EQUAL: 0xd9,
-	BINOP_LESS_EQUAL: 0xda,
-	BINOP_MORE_EQUAL: 0xdb,
-	BINOP_NOT_EQUAL: 0xdc,
-	BINOP_IN: 0xdd,
-	BINOP_IS: 0xde,
-	BINOP_EXCEPTION_MATCH: 0xdf,
-	BINOP_INPLACE_OR: 0xe0,
-	BINOP_INPLACE_XOR: 0xe1,
-	BINOP_INPLACE_AND: 0xe2,
-	BINOP_INPLACE_LSHIFT: 0xe3,
-	BINOP_INPLACE_RSHIFT: 0xe4,
-	BINOP_INPLACE_ADD: 0xe5,
-	BINOP_INPLACE_SUBTRACT: 0xe6,
-	BINOP_INPLACE_MULTIPLY: 0xe7,
-	BINOP_INPLACE_MAT_MULTIPLY: 0xe8,
-	BINOP_INPLACE_FLOOR_DIVIDE: 0xe9,
-	BINOP_INPLACE_TRUE_DIVIDE: 0xea,
-	BINOP_INPLACE_MODULO: 0xeb,
-	BINOP_INPLACE_POWER: 0xec,
-	BINOP_OR: 0xed,
-	BINOP_XOR: 0xee,
-	BINOP_AND: 0xef,
-	BINOP_LSHIFT: 0xf0,
-	BINOP_RSHIFT: 0xf1,
-	BINOP_ADD: 0xf2,
-	BINOP_SUBTRACT: 0xf3,
-	BINOP_MULTIPLY: 0xf4,
-	BINOP_MAT_MULTIPLY: 0xf5,
-	BINOP_FLOOR_DIVIDE: 0xf6,
-	BINOP_TRUE_DIVIDE: 0xf7,
-	BINOP_MODULO: 0xf8,
-	BINOP_POWER: 0xf9,
-	BINOP_DIVMOD: 0xfa,
-	BINOP_CONTAINS: 0xfb,
-	BINOP_REVERSE_OR: 0xfc,
-	BINOP_REVERSE_XOR: 0xfd,
-	BINOP_REVERSE_AND: 0xfe,
-	BINOP_REVERSE_LSHIFT: 0xff,
-	BINOP_REVERSE_RSHIFT: 0x100,
-	BINOP_REVERSE_ADD: 0x101,
-	BINOP_REVERSE_SUBTRACT: 0x102,
-	BINOP_REVERSE_MULTIPLY: 0x103,
-	BINOP_REVERSE_MAT_MULTIPLY: 0x104,
-	BINOP_REVERSE_FLOOR_DIVIDE: 0x105,
-	BINOP_REVERSE_TRUE_DIVIDE: 0x106,
-	BINOP_REVERSE_MODULO: 0x107,
-	BINOP_REVERSE_POWER: 0x108,
-};
-let opsnm = {
-	MP_BC_LOAD_CONST_FALSE: "false",
-	MP_BC_LOAD_CONST_NONE: "none",
-	MP_BC_LOAD_CONST_TRUE: "true",
-	MP_BC_LOAD_CONST_SMALL_INT: "int",
-	MP_BC_LOAD_CONST_STRING: "str",
-	MP_BC_LOAD_CONST_OBJ: "constobj",
-	MP_BC_LOAD_NULL: "null",
-	MP_BC_LOAD_FAST_N: "MP_BC_LOAD_FAST_N",
-	MP_BC_LOAD_DEREF: "deref",
-	MP_BC_LOAD_NAME: "loadname",
-	MP_BC_LOAD_GLOBAL: "glbl",
-	MP_BC_LOAD_ATTR: "attr",
-	MP_BC_LOAD_METHOD: "method",
-	MP_BC_LOAD_SUPER_METHOD: "supermethod",
-	MP_BC_LOAD_BUILD_CLASS: "buildclass",
-	MP_BC_LOAD_SUBSCR: "ld.arrsub",
-	MP_BC_STORE_FAST_N: "MP_BC_STORE_FAST_N",
-	MP_BC_STORE_DEREF: "st.deref",
-	MP_BC_STORE_NAME: "st.name",
-	MP_BC_STORE_GLOBAL: "st.glbl",
-	MP_BC_STORE_ATTR: "st.attr",
-	MP_BC_STORE_SUBSCR: "st.arrsub",
-	MP_BC_DELETE_FAST: "del.fast",
-	MP_BC_DELETE_DEREF: "del.deref",
-	MP_BC_DELETE_NAME: "del.name",
-	MP_BC_DELETE_GLOBAL: "del.glbl",
-	MP_BC_DUP_TOP: "dup",
-	MP_BC_DUP_TOP_TWO: "dup2",
-	MP_BC_POP_TOP: "pop",
-	MP_BC_ROT_TWO: "rot",
-	MP_BC_ROT_THREE: "rot3",
-	MP_BC_JUMP: "jmp",
-	MP_BC_POP_JUMP_IF_TRUE: "btrue",
-	MP_BC_POP_JUMP_IF_FALSE: "bfalse",
-	MP_BC_JUMP_IF_TRUE_OR_POP: "btrue.pop",
-	MP_BC_JUMP_IF_FALSE_OR_POP: "bfalse.pop",
-	MP_BC_UNWIND_JUMP: "jmp.unwind",
-	MP_BC_SETUP_WITH: "with.setup",
-	MP_BC_SETUP_EXCEPT: "except.setup",
-	MP_BC_SETUP_FINALLY: "finally.setup",
-	MP_BC_POP_EXCEPT_JUMP: "except.jump",
-	MP_BC_FOR_ITER: "for.iter",
-	MP_BC_WITH_CLEANUP: "with.chleanup",
-	MP_BC_END_FINALLY: "finally.end",
-	MP_BC_GET_ITER: "ld.iter",
-	MP_BC_GET_ITER_STACK: "ld.iterstack",
-	MP_BC_BUILD_TUPLE: "tuple",
-	MP_BC_BUILD_LIST: "list",
-	MP_BC_BUILD_MAP: "map",
-	MP_BC_STORE_MAP: "st.map",
-	MP_BC_BUILD_SET: "MP_BC_BUILD_SET",
-	MP_BC_BUILD_SLICE: "slice",
-	MP_BC_STORE_COMP: "MP_BC_STORE_COMP",
-	MP_BC_UNPACK_SEQUENCE: "MP_BC_UNPACK_SEQUENCE",
-	MP_BC_UNPACK_EX: "MP_BC_UNPACK_EX",
-	MP_BC_RETURN_VALUE: "ret",
-	MP_BC_RAISE_LAST: "rethrow",
-	MP_BC_RAISE_OBJ: "raiseobj",
-	MP_BC_RAISE_FROM: "throwfrom",
-	MP_BC_YIELD_VALUE: "yield",
-	MP_BC_YIELD_FROM: "yieldfrom",
-	MP_BC_MAKE_FUNCTION: "mkfun",
-	MP_BC_MAKE_FUNCTION_DEFARGS: "mkfun.defargs",
-	MP_BC_MAKE_CLOSURE: "mkclosure",
-	MP_BC_MAKE_CLOSURE_DEFARGS: "mkclosure.defargs",
-	MP_BC_CALL_FUNCTION: "call",
-	MP_BC_CALL_FUNCTION_VAR_KW: "call.kw",
-	MP_BC_CALL_METHOD: "call.meth",
-	MP_BC_CALL_METHOD_VAR_KW: "call.kvmeth",
-	MP_BC_IMPORT_NAME: "import.nm",
-	MP_BC_IMPORT_FROM: "import.from",
-	MP_BC_IMPORT_STAR: "import.all",
-
-	UNOP_POSITIVE: "assertint",
-	UNOP_NEGATIVE: "neg",
-	UNOP_INVERT: "invert",
-	UNOP_NOT: "not",
-
-	BINOP_LESS: "lt",
-	BINOP_MORE: "qt",
-	BINOP_EQUAL: "eq",
-	BINOP_LESS_EQUAL: "le",
-	BINOP_MORE_EQUAL: "qe",
-	BINOP_NOT_EQUAL: "ne",
-	BINOP_IN: "in",
-	BINOP_IS: "is",
-	BINOP_EXCEPTION_MATCH: "ematch",
-	BINOP_INPLACE_OR: "or.in",
-	BINOP_INPLACE_XOR: "xor.in",
-	BINOP_INPLACE_AND: "and.in",
-	BINOP_INPLACE_LSHIFT: "shl.in",
-	BINOP_INPLACE_RSHIFT: "shr.in",
-	BINOP_INPLACE_ADD: "add.in",
-	BINOP_INPLACE_SUBTRACT: "sub.in",
-	BINOP_INPLACE_MULTIPLY: "mul.in",
-	BINOP_INPLACE_MAT_MULTIPLY: "matmul.in",
-	BINOP_INPLACE_FLOOR_DIVIDE: "floordiv.in",
-	BINOP_INPLACE_TRUE_DIVIDE: "div.in",
-	BINOP_INPLACE_MODULO: "mod.in",
-	BINOP_INPLACE_POWER: "pow.in",
-	BINOP_OR: "or",
-	BINOP_XOR: "xor",
-	BINOP_AND: "and",
-	BINOP_LSHIFT: "shl",
-	BINOP_RSHIFT: "shr",
-	BINOP_ADD: "add",
-	BINOP_SUBTRACT: "sub",
-	BINOP_MULTIPLY: "mul",
-	BINOP_MAT_MULTIPLY: "matmul",
-	BINOP_FLOOR_DIVIDE: "floordiv",
-	BINOP_TRUE_DIVIDE: "truediv",
-	BINOP_MODULO: "mod",
-	BINOP_POWER: "pow",
-	BINOP_DIVMOD: "divmod",
-	BINOP_CONTAINS: "in",
-	BINOP_REVERSE_OR: "or.rev",
-	BINOP_REVERSE_XOR: "xor.rev",
-	BINOP_REVERSE_AND: "and.rev",
-	BINOP_REVERSE_LSHIFT: "shl.rev",
-	BINOP_REVERSE_RSHIFT: "shr.rev",
-	BINOP_REVERSE_ADD: "add.rev",
-	BINOP_REVERSE_SUBTRACT: "sub.rev",
-	BINOP_REVERSE_MULTIPLY: "mul.rev",
-	BINOP_REVERSE_MAT_MULTIPLY: "matmul.rev",
-	BINOP_REVERSE_FLOOR_DIVIDE: "floordiv.rev",
-	BINOP_REVERSE_TRUE_DIVIDE: "truediv.rev",
-	BINOP_REVERSE_MODULO: "mod.rev",
-	BINOP_REVERSE_POWER: "pow.rev",
-};
-let optypes = {
-	[importmap.MP_BC_LOAD_CONST_FALSE]: "MP_BC_LOAD_CONST_FALSE",
-	[importmap.MP_BC_LOAD_CONST_NONE]: "MP_BC_LOAD_CONST_NONE",
-	[importmap.MP_BC_LOAD_CONST_TRUE]: "MP_BC_LOAD_CONST_TRUE",
-	[importmap.MP_BC_LOAD_CONST_SMALL_INT]: "MP_BC_LOAD_CONST_SMALL_INT",
-	[importmap.MP_BC_LOAD_CONST_STRING]: "MP_BC_LOAD_CONST_STRING",
-	[importmap.MP_BC_LOAD_CONST_OBJ]: "MP_BC_LOAD_CONST_OBJ",
-	[importmap.MP_BC_LOAD_NULL]: "MP_BC_LOAD_NULL",
-	[importmap.MP_BC_LOAD_FAST_N]: "MP_BC_LOAD_FAST_N",
-	[importmap.MP_BC_LOAD_DEREF]: "MP_BC_LOAD_DEREF",
-	[importmap.MP_BC_LOAD_NAME]: "MP_BC_LOAD_NAME",
-	[importmap.MP_BC_LOAD_GLOBAL]: "MP_BC_LOAD_GLOBAL",
-	[importmap.MP_BC_LOAD_ATTR]: "MP_BC_LOAD_ATTR",
-	[importmap.MP_BC_LOAD_METHOD]: "MP_BC_LOAD_METHOD",
-	[importmap.MP_BC_LOAD_SUPER_METHOD]: "MP_BC_LOAD_SUPER_METHOD",
-	[importmap.MP_BC_LOAD_BUILD_CLASS]: "MP_BC_LOAD_BUILD_CLASS",
-	[importmap.MP_BC_LOAD_SUBSCR]: "MP_BC_LOAD_SUBSCR",
-	[importmap.MP_BC_STORE_FAST_N]: "MP_BC_STORE_FAST_N",
-	[importmap.MP_BC_STORE_DEREF]: "MP_BC_STORE_DEREF",
-	[importmap.MP_BC_STORE_NAME]: "MP_BC_STORE_NAME",
-	[importmap.MP_BC_STORE_GLOBAL]: "MP_BC_STORE_GLOBAL",
-	[importmap.MP_BC_STORE_ATTR]: "MP_BC_STORE_ATTR",
-	[importmap.MP_BC_STORE_SUBSCR]: "MP_BC_STORE_SUBSCR",
-	[importmap.MP_BC_DELETE_FAST]: "MP_BC_DELETE_FAST",
-	[importmap.MP_BC_DELETE_DEREF]: "MP_BC_DELETE_DEREF",
-	[importmap.MP_BC_DELETE_NAME]: "MP_BC_DELETE_NAME",
-	[importmap.MP_BC_DELETE_GLOBAL]: "MP_BC_DELETE_GLOBAL",
-	[importmap.MP_BC_DUP_TOP]: "MP_BC_DUP_TOP",
-	[importmap.MP_BC_DUP_TOP_TWO]: "MP_BC_DUP_TOP_TWO",
-	[importmap.MP_BC_POP_TOP]: "MP_BC_POP_TOP",
-	[importmap.MP_BC_ROT_TWO]: "MP_BC_ROT_TWO",
-	[importmap.MP_BC_ROT_THREE]: "MP_BC_ROT_THREE",
-	[importmap.MP_BC_JUMP]: "MP_BC_JUMP",
-	[importmap.MP_BC_POP_JUMP_IF_TRUE]: "MP_BC_POP_JUMP_IF_TRUE",
-	[importmap.MP_BC_POP_JUMP_IF_FALSE]: "MP_BC_POP_JUMP_IF_FALSE",
-	[importmap.MP_BC_JUMP_IF_TRUE_OR_POP]: "MP_BC_JUMP_IF_TRUE_OR_POP",
-	[importmap.MP_BC_JUMP_IF_FALSE_OR_POP]: "MP_BC_JUMP_IF_FALSE_OR_POP",
-	[importmap.MP_BC_UNWIND_JUMP]: "MP_BC_UNWIND_JUMP",
-	[importmap.MP_BC_SETUP_WITH]: "MP_BC_SETUP_WITH",
-	[importmap.MP_BC_SETUP_EXCEPT]: "MP_BC_SETUP_EXCEPT",
-	[importmap.MP_BC_SETUP_FINALLY]: "MP_BC_SETUP_FINALLY",
-	[importmap.MP_BC_POP_EXCEPT_JUMP]: "MP_BC_POP_EXCEPT_JUMP",
-	[importmap.MP_BC_FOR_ITER]: "MP_BC_FOR_ITER",
-	[importmap.MP_BC_WITH_CLEANUP]: "MP_BC_WITH_CLEANUP",
-	[importmap.MP_BC_END_FINALLY]: "MP_BC_END_FINALLY",
-	[importmap.MP_BC_GET_ITER]: "MP_BC_GET_ITER",
-	[importmap.MP_BC_GET_ITER_STACK]: "MP_BC_GET_ITER_STACK",
-	[importmap.MP_BC_BUILD_TUPLE]: "MP_BC_BUILD_TUPLE",
-	[importmap.MP_BC_BUILD_LIST]: "MP_BC_BUILD_LIST",
-	[importmap.MP_BC_BUILD_MAP]: "MP_BC_BUILD_MAP",
-	[importmap.MP_BC_STORE_MAP]: "MP_BC_STORE_MAP",
-	[importmap.MP_BC_BUILD_SET]: "MP_BC_BUILD_SET",
-	[importmap.MP_BC_BUILD_SLICE]: "MP_BC_BUILD_SLICE",
-	[importmap.MP_BC_STORE_COMP]: "MP_BC_STORE_COMP",
-	[importmap.MP_BC_UNPACK_SEQUENCE]: "MP_BC_UNPACK_SEQUENCE",
-	[importmap.MP_BC_UNPACK_EX]: "MP_BC_UNPACK_EX",
-	[importmap.MP_BC_RETURN_VALUE]: "MP_BC_RETURN_VALUE",
-	[importmap.MP_BC_RAISE_LAST]: "MP_BC_RAISE_LAST",
-	[importmap.MP_BC_RAISE_OBJ]: "MP_BC_RAISE_OBJ",
-	[importmap.MP_BC_RAISE_FROM]: "MP_BC_RAISE_FROM",
-	[importmap.MP_BC_YIELD_VALUE]: "MP_BC_YIELD_VALUE",
-	[importmap.MP_BC_YIELD_FROM]: "MP_BC_YIELD_FROM",
-	[importmap.MP_BC_MAKE_FUNCTION]: "MP_BC_MAKE_FUNCTION",
-	[importmap.MP_BC_MAKE_FUNCTION_DEFARGS]: "MP_BC_MAKE_FUNCTION_DEFARGS",
-	[importmap.MP_BC_MAKE_CLOSURE]: "MP_BC_MAKE_CLOSURE",
-	[importmap.MP_BC_MAKE_CLOSURE_DEFARGS]: "MP_BC_MAKE_CLOSURE_DEFARGS",
-	[importmap.MP_BC_CALL_FUNCTION]: "MP_BC_CALL_FUNCTION",
-	[importmap.MP_BC_CALL_FUNCTION_VAR_KW]: "MP_BC_CALL_FUNCTION_VAR_KW",
-	[importmap.MP_BC_CALL_METHOD]: "MP_BC_CALL_METHOD",
-	[importmap.MP_BC_CALL_METHOD_VAR_KW]: "MP_BC_CALL_METHOD_VAR_KW",
-	[importmap.MP_BC_IMPORT_NAME]: "MP_BC_IMPORT_NAME",
-	[importmap.MP_BC_IMPORT_FROM]: "MP_BC_IMPORT_FROM",
-	[importmap.MP_BC_IMPORT_STAR]: "MP_BC_IMPORT_STAR",
-
-	[importmap.UNOP_POSITIVE]: "UNOP_POSITIVE",
-	[importmap.UNOP_NEGATIVE]: "UNOP_NEGATIVE",
-	[importmap.UNOP_INVERT]: "UNOP_INVERT",
-	[importmap.UNOP_NOT]: "UNOP_NOT",
-
-	[importmap.BINOP_LESS]: "BINOP_LESS",
-	[importmap.BINOP_MORE]: "BINOP_MORE",
-	[importmap.BINOP_EQUAL]: "BINOP_EQUAL",
-	[importmap.BINOP_LESS_EQUAL]: "BINOP_LESS_EQUAL",
-	[importmap.BINOP_MORE_EQUAL]: "BINOP_MORE_EQUAL",
-	[importmap.BINOP_NOT_EQUAL]: "BINOP_NOT_EQUAL",
-	[importmap.BINOP_IN]: "BINOP_IN",
-	[importmap.BINOP_IS]: "BINOP_IS",
-	[importmap.BINOP_EXCEPTION_MATCH]: "BINOP_EXCEPTION_MATCH",
-	[importmap.BINOP_INPLACE_OR]: "BINOP_INPLACE_OR",
-	[importmap.BINOP_INPLACE_XOR]: "BINOP_INPLACE_XOR",
-	[importmap.BINOP_INPLACE_AND]: "BINOP_INPLACE_AND",
-	[importmap.BINOP_INPLACE_LSHIFT]: "BINOP_INPLACE_LSHIFT",
-	[importmap.BINOP_INPLACE_RSHIFT]: "BINOP_INPLACE_RSHIFT",
-	[importmap.BINOP_INPLACE_ADD]: "BINOP_INPLACE_ADD",
-	[importmap.BINOP_INPLACE_SUBTRACT]: "BINOP_INPLACE_SUBTRACT",
-	[importmap.BINOP_INPLACE_MULTIPLY]: "BINOP_INPLACE_MULTIPLY",
-	[importmap.BINOP_INPLACE_MAT_MULTIPLY]: "BINOP_INPLACE_MAT_MULTIPLY",
-	[importmap.BINOP_INPLACE_FLOOR_DIVIDE]: "BINOP_INPLACE_FLOOR_DIVIDE",
-	[importmap.BINOP_INPLACE_TRUE_DIVIDE]: "BINOP_INPLACE_TRUE_DIVIDE",
-	[importmap.BINOP_INPLACE_MODULO]: "BINOP_INPLACE_MODULO",
-	[importmap.BINOP_INPLACE_POWER]: "BINOP_INPLACE_POWER",
-	[importmap.BINOP_OR]: "BINOP_OR",
-	[importmap.BINOP_XOR]: "BINOP_XOR",
-	[importmap.BINOP_AND]: "BINOP_AND",
-	[importmap.BINOP_LSHIFT]: "BINOP_LSHIFT",
-	[importmap.BINOP_RSHIFT]: "BINOP_RSHIFT",
-	[importmap.BINOP_ADD]: "BINOP_ADD",
-	[importmap.BINOP_SUBTRACT]: "BINOP_SUBTRACT",
-	[importmap.BINOP_MULTIPLY]: "BINOP_MULTIPLY",
-	[importmap.BINOP_MAT_MULTIPLY]: "BINOP_MAT_MULTIPLY",
-	[importmap.BINOP_FLOOR_DIVIDE]: "BINOP_FLOOR_DIVIDE",
-	[importmap.BINOP_TRUE_DIVIDE]: "BINOP_TRUE_DIVIDE",
-	[importmap.BINOP_MODULO]: "BINOP_MODULO",
-	[importmap.BINOP_POWER]: "BINOP_POWER",
-	[importmap.BINOP_DIVMOD]: "BINOP_DIVMOD",
-	[importmap.BINOP_CONTAINS]: "BINOP_CONTAINS",
-	[importmap.BINOP_REVERSE_OR]: "BINOP_REVERSE_OR",
-	[importmap.BINOP_REVERSE_XOR]: "BINOP_REVERSE_XOR",
-	[importmap.BINOP_REVERSE_AND]: "BINOP_REVERSE_AND",
-	[importmap.BINOP_REVERSE_LSHIFT]: "BINOP_REVERSE_LSHIFT",
-	[importmap.BINOP_REVERSE_RSHIFT]: "BINOP_REVERSE_RSHIFT",
-	[importmap.BINOP_REVERSE_ADD]: "BINOP_REVERSE_ADD",
-	[importmap.BINOP_REVERSE_SUBTRACT]: "BINOP_REVERSE_SUBTRACT",
-	[importmap.BINOP_REVERSE_MULTIPLY]: "BINOP_REVERSE_MULTIPLY",
-	[importmap.BINOP_REVERSE_MAT_MULTIPLY]: "BINOP_REVERSE_MAT_MULTIPLY",
-	[importmap.BINOP_REVERSE_FLOOR_DIVIDE]: "BINOP_REVERSE_FLOOR_DIVIDE",
-	[importmap.BINOP_REVERSE_TRUE_DIVIDE]: "BINOP_REVERSE_TRUE_DIVIDE",
-	[importmap.BINOP_REVERSE_MODULO]: "BINOP_REVERSE_MODULO",
-	[importmap.BINOP_REVERSE_POWER]: "BINOP_REVERSE_POWER",
-};
-
-let branches = ["btrue", "bfalse", "btrue.pop", "bfalse.pop", "jmp.unwind", "jmp"];
-let outg = [];
-function out(s: string) {
-	outg.push(s);
-}
-function disassemblefrom_c_opinput(opinput0: string[], p: { [key: string]: string }) {
-	out(p[opinput0[1].split("& 0xff")[0].trim()] + ":");
-	let off = 0;
-	let usedAdderesses = new Set<number>();
-	let opinput = opinput0.map((e) => {
-		let x = e
-			.trim()
-			.split(",")
-			.map((e) => e.trim())
-			.filter((e) => e);
-		off += x.length;
-		x.push(off.toString());
-		return x
-			.map((e) => (e.startsWith("0x") ? +e : e.endsWith(">> 8") ? "" : e.endsWith("& 0xff") ? p[e.split(" ")[0]] : e))
-			.filter((e) => e !== "")
-			.map((e, i) =>
-				typeof e == "number" && i == 0
-					? opsnm[optypes[e]] ||
-					  (typeof e == "number" && e >= 0xb0 && e < 0xc0 && i == 0
-							? `ldloc.${e - 0xb0}`
-							: typeof e == "number" && e >= 0xc0 && e < 0xd0 && i == 0
-							? `stloc.${e - 0xc0}`
-							: typeof e == "number" && e >= 0x80 && e < 0x90 && i == 0
-							? `int ${e - 0x80}`
-							: e)
-					: e
-			)
-			.map((e, i, a) =>
-				branches.includes(a[0]) ? (usedAdderesses.add((a[1] | (a[2] << 8)) - 0x8000 + off), [[e, ".L" + ((a[1] | (a[2] << 8)) - 0x8000 + off).toString(), [], e][i]].flat()) : [e]
-			)
-			.flat();
-	});
-	off = 0;
-	let opinput1 = opinput.slice(4).map((e, i, a) => {
-		if (i >= 1 && usedAdderesses.has(+a[i - 1][a[i - 1].length - 1])) {
-			return (
-				".L" +
-				a[i - 1][a[i - 1].length - 1] +
-				":\n\t" +
-				e
-					.slice(0, -1)
-					.map((e, i) => (i > 1 ? `, ${e}` : e))
-					.join(" ")
-					.replaceAll(" , ", ", ")
-			);
+function evalopwithstack(shadow: string[], i: number, x: fmt, stack: string[], tablevel: number): { stack: string[]; tablevel: number } {
+	let e = shadow[i]
+	if (e == null) {
+		for (let e of stack) {
+			console.log('    '.repeat(tablevel) + '$push(' + e + ')')
 		}
-		return (
-			"\t" +
-			e
-				.slice(0, -1)
-				.map((e, i) => (i > 1 ? `, ${e}` : e))
-				.join(" ")
-				.replaceAll(" , ", ", ")
-		);
-	});
-	out(opinput1.join("\n"));
-}
-
-// size_t slab = (ip[0] | (ip[1] << 8)) - 0x8000; ip += 2
-
-// function pattern(
-//   p: TemplateStringsArray,
-//   ...s: ((p: string[]) => string | string[])[]
-// ) {
-//   let all = String.raw(p, s.map((_e, i) => `%%__fn${i}%%`)).trim().split("\n")
-//     .map((e) => e.split(";")[0].trim()).filter((e) => e).map((e) =>
-//       e.split(" ")
-//     );
-//   let ops = new Map<string, (p: string[]) => string | string[]>();
-//   let dat = new Map<string, string[]>();
-//   let output = [];
-//   let fork = [...opinput];
-//   for (let insn of all) {
-//     let opcode = insn[0];
-//     if (opcode == "%op") {
-//       ops.set(insn[1], s.shift());
-//     }
-//     if (opcode == "%match") {
-//       let dat = insn.slice(1);
-//     }
-//     if (opcode == "%commit") {
-//       opinput = [...fork];
-//     }
-//   }
-// }
-// // roll in glbl/attr into push
-// pattern`
-// %op do_dot_join ${(p) => p.join(".")}
-// %match glbl $id
-// %rep %match attr $id2
-// %produce push %do_dot_join($id $id2)
-// %commit`;
-
-for (let type of execSync("find in -type f")
-	.toString()
-	.trim()
-	.split("\n")
-	.map((e) => ({ in: e, out: "out/" + e.slice(3).split(".")[0] + ".s" }))) {
-	if (!type.in.endsWith(".mpy")) continue;
-	let outdotc = execSync("python micropython/tools/mpy-tool.py --freeze " + type.in).toString();
-	console.log(type);
-	let keys = outdotc
-		.split("enum {")[1]
-		.split("}")[0]
-		.trim()
-		.split("\n")
-		.map((e) => e.trim().split(",")[0]);
-	let values = outdotc
-		.split("mp_qstr_frozen_const_pool")[1]
-		.split("{")[2]
-		.split("}")[0]
-		.trim()
-		.split("\n")
-		.map((e) => e.split('"')[3]);
-	let entries = keys.map((e, i) => [e, values[i]]);
-	let djson = Object.fromEntries(entries);
-	let osn = outdotc.split("\n");
-	let data = osn.map((e, i) => [e, i] as const).filter((e) => e[0].startsWith("STATIC const byte fun_data_"));
-	outg = [];
-	data.map((e) =>
-		osn
-			.slice(e[1] + 1)
-			.join("\n")
-			.split("};")[0]
-			.split("\n")
-			.map((e) => e.trim())
-	).forEach((t) => {
-		disassemblefrom_c_opinput(t, djson);
-	});
-	if (!existsSync(type.out.split("/").slice(0, -1).join("/"))) {
-		execSync("mkdir -p " + type.out.split("/").slice(0, -1).join("/"));
+		stack = []
+		console.log('    '.repeat(tablevel) + 'BROKEN_OP`' + x[i].join(' ') + '`')
+	} else {
+		let cnt = Math.max(-1, ...(e.match(/\$pop\([0-9]+\)/g) || []).map(e => +e.slice(5, -1))) + 1
+		let data = []
+		for (let i = 0; i < cnt; i++) {
+			if (stack.length) {
+				data.push(stack.pop())
+			} else {
+				data.push('__pop')
+			}
+		}
+		let out = e
+			.replaceAll(/\$pop\([0-9]+\)/g, p => data[p.slice(5, -1)])
+			.replaceAll(/\$fun\([^\)]+\)/g, p => {
+				console.log('    '.repeat(tablevel) + 'def ' + p.slice(5, -1) + '(*args):')
+				decompile_text(grps.get(p.slice(5, -1)), tablevel + 1) // '__tfn_' + p
+				return p.slice(5, -1)
+			})
+			.replaceAll(/\$fst\([^,]+,[^\)]+\)/g, p => {
+				let [id, nm] = p.slice(5, -1).split(',')
+				console.log('    '.repeat(tablevel) + 'def ' + nm + '(*args):')
+				decompile_text(grps.get(id), tablevel + 1) // '__tfn_' + p
+				return '$nop'
+			})
+			.replaceAll(/\$warp\([0-9]+\)/g, p => {
+				let q = +p.slice(6, -1)
+				//evalopwithstack
+				// (shadow: string[], i: number, x: fmt, stack: string[], tablevel: number)
+				let new_stk = []
+				for (let li = q; x[li][0] != 'btrue'; li++) {
+					let xu = evalopwithstack(shadow, li, x, new_stk, 10)
+					new_stk = xu.stack
+				}
+				return new_stk[0]
+			})
+			.replaceAll(/\$ign\([^\)]+\)/g, '')
+		if (out == '$nop') return { stack, tablevel }
+		let tableveln = tablevel
+		if (out.endsWith('{')) {
+			tableveln++
+			out = out.slice(0, -1).trim() + ':'
+		}
+		if (out.startsWith('}')) {
+			tableveln--
+			if (out != '}') tablevel--
+			out = out.slice(1).trim()
+		}
+		if (out.startsWith('$push(') && out.endsWith(')')) {
+			stack.push(out.slice(6, -1))
+		} else if (out.startsWith('$psh2(') && out.endsWith(')')) {
+			stack.push(out.slice(6, -1))
+			stack.push(out.slice(6, -1))
+		} else if (out == '$dup') {
+			let x = stack.pop()
+			stack.push(x)
+			stack.push(x)
+		} else if (out == '}') {
+			console.log('    '.repeat(tablevel) + stack.pop())
+		} else if (out == '#') {
+		} else {
+			console.log('    '.repeat(tablevel) + out)
+		}
+		tablevel = tableveln
 	}
-	writeFileSync(type.out, outg.join("\n"));
+	return { stack, tablevel }
 }
+
+// 	ret ; return $pop()
+let grps = new Map<string, string>()
+let ids = []
+function decompile_text(text: string, tlba: number = 0) {
+	let map_z = new Map<string, string>()
+	let x2: fmt = text
+		.split('\n')
+		.slice(1)
+		.map((e, i) => (e.startsWith('\t') ? e.trim().split(' ') : (map_z.set(e.slice(0, -1), (+i + 1).toString()), ['nop'])))
+		.map(e => e.map(f => map_z.get(f) || f))
+
+	let x: fmt = x2
+	let decompilerShadow: (string | null)[] = x.map(_ => null)
+	let branches = ['btrue', 'bfalse', 'btrue.pop', 'bfalse.pop', 'jmp.unwind', 'jmp', 'for.iter']
+
+	let opcode2data = new Map(
+		Object.entries({
+			true: '$push(True)',
+			false: '$push(False)',
+			pop: '$pop(0)',
+			none: '$push(None)',
+			ret: 'return $pop(0)',
+			'st.name': '$1 = $pop(0)',
+			loadname: '$push($1)',
+			glbl: '$push($1)',
+			int: '$push($1)',
+			deref: '$push(args[$1])',
+			'st.deref': 'args[$1] = $pop(0)',
+
+			'ldloc.0': '$push(args[0])',
+			'ldloc.1': '$push(args[1])',
+			'ldloc.2': '$push(args[2])',
+			'ldloc.3': '$push(args[3])',
+			'ldloc.4': '$push(args[4])',
+			'ldloc.5': '$push(args[5])',
+			'ldloc.6': '$push(args[6])',
+			'ldloc.7': '$push(args[7])',
+			'ldloc.8': '$push(args[8])',
+			'ldloc.9': '$push(args[9])',
+			'ldloc.10': '$push(args[10])',
+
+			'stloc.0': 'args[0] = $pop(0)',
+			'stloc.1': 'args[1] = $pop(0)',
+			'stloc.2': 'args[2] = $pop(0)',
+			'stloc.3': 'args[3] = $pop(0)',
+			'stloc.4': 'args[4] = $pop(0)',
+			'stloc.5': 'args[5] = $pop(0)',
+			'stloc.6': 'args[6] = $pop(0)',
+			'stloc.7': 'args[7] = $pop(0)',
+			'stloc.8': 'args[8] = $pop(0)',
+			'stloc.9': 'args[9] = $pop(0)',
+			'stloc.10': 'args[10] = $pop(0)',
+
+			add: '$push($pop(1) + $pop(0))',
+			lt: '$push($pop(1) < $pop(0))',
+			'sub.in': '$push($pop(1) - $pop(0))',
+
+			raiseobj: 'raise $pop(0)',
+			'del.name': 'del $1',
+			nop: '$nop',
+			buildclass: '$push(__build_class__)',
+			'st.attr': '$pop(0).$1 = $pop(1)',
+			attr: '$push($pop(0).$1)',
+			method: '$push($pop(0).$1)',
+			null: 'null',
+			'ld.arrsub': '$push($pop(1)[$pop(0)])',
+			is: '$push($pop(1) is $pop(0))',
+			ematch: '$push($pop(1) is $pop(0))',
+			not: `$push(not $pop(0))`,
+			dup: `$dup`,
+		})
+	)
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (opcode2data.has(op[0])) {
+			decompilerShadow[opidx] = opcode2data.get(op[0]).replace(/\$[0-9]/g, p => op[p.slice(1)])
+		}
+	}
+
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'str') {
+			// if case
+			// op[1]
+			decompilerShadow[opidx] = `$push(${JSON.stringify(op[1])})`
+		}
+	}
+
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'except.setup') {
+			let current = +op[1]
+			decompilerShadow[opidx] = 'try {'
+			decompilerShadow[current - 2] = `$push(__except)`
+			decompilerShadow[current - 1] = `} except Exception as __except {`
+			let end_ptr = +x[current - 2][1] - 2
+			do {
+				decompilerShadow[current + 5] = `$nop`
+				let offset = +x[current + 5][1]
+				for (let i = 0; i < 5; i++) decompilerShadow[offset + i] = '$nop'
+				decompilerShadow[offset + 5] = '$psh2(__except)'
+				current = +x[current + 3][1]
+			} while (current != end_ptr)
+			decompilerShadow[end_ptr - 3] = '}'
+		}
+	}
+
+	for (let opidx in x) {
+		let op = x[opidx]
+		let nxop = x[1 + +opidx]
+		if (op[0] == 'import.nm') {
+			let my_name = op[1]
+			if (nxop[0] == 'st.name') {
+				decompilerShadow[1 + +opidx] = `$nop`
+				let used_name = nxop[1]
+				if (used_name != my_name) {
+					decompilerShadow[+opidx] = `import ${my_name} as ${used_name} $ign($pop(1))`
+				} else {
+					decompilerShadow[+opidx] = `import ${my_name} $ign($pop(1))`
+				}
+			}
+			if (nxop[0] == 'import.all') {
+				decompilerShadow[1 + +opidx] = `$nop`
+				decompilerShadow[+opidx] = `from ${my_name} import * $ign($pop(1))`
+			}
+			if (nxop[0] == 'import.from') {
+				// we need to iterate :(
+				// that's the only way.
+				let map = []
+				for (let i = +opidx + 1; x[i][0] != 'pop'; i += 2, decompilerShadow[i] = '$nop') {
+					let [mine, theirs] = [x[i][1], x[i + 1][1]]
+					if (mine == theirs) {
+						map.push(mine)
+					} else {
+						map.push(`${mine} as ${theirs}`)
+					}
+					decompilerShadow[i] = `$nop`
+					decompilerShadow[i + 1] = `$nop`
+				}
+				decompilerShadow[opidx] = `from ${my_name} import ${map.join(', ')} $ign($pop(1))`
+			}
+		}
+	}
+	// while-loop
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'jmp') {
+			let tgop = +op[1]
+			for (let i = +op[1]; i < x.length && !branches.includes(x[i - 1][0]); i++) {
+				if (x[i][0] == 'btrue') {
+					tgop = i
+				}
+			}
+			if (x[tgop] && x[tgop][0] == 'btrue') {
+				decompilerShadow[opidx] = `while $warp(${op[1]}) {`
+				decompilerShadow[tgop] = '}'
+			}
+		}
+	}
+	// for-loop
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'ld.iterstack' && x[+opidx + 2][0] == 'for.iter' && x[+opidx + 1][0] == 'nop' && x[+opidx + 3][0] == 'st.name') {
+			decompilerShadow[opidx] = '$nop'
+			decompilerShadow[+opidx + 2] = '$nop'
+			decompilerShadow[+opidx + 3] = `for ${x[+opidx + 3][1]} in $pop(0) {`
+			decompilerShadow[+x[+opidx + 2][1] - 2] = '}'
+			// we land insn after }
+			// on the jmp
+		}
+	}
+	// fun-norm
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'mkfun') {
+			// call
+			// $push($pop(2)($pop(0), $pop(1)))
+			decompilerShadow[opidx] = `$push($fun(${op[1]}))`
+		}
+		if (op[0] == 'mkfun.defargs') {
+			// call
+			// $push($pop(2)($pop(0), $pop(1)))
+			decompilerShadow[opidx] = `$push($fun(${op[1]})$ign($pop(0))`
+		}
+	}
+	// fun-store
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'mkfun' && x[1 + +opidx][0] == 'st.name') {
+			decompilerShadow[+opidx] = `$fst(${op[1]},${x[1 + +opidx][1]})`
+			decompilerShadow[+opidx + 1] = `$nop`
+		}
+	}
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'bfalse' && +op[1] > +opidx) {
+			// if case
+			// op[1]
+			// this is actually a pop right after the if case to pop off the argument to bfalse.
+			if (x[+op[1] - 2][0] == 'jmp') {
+				decompilerShadow[opidx] = 'if $pop(0) {'
+				decompilerShadow[+op[1] - 2] = '} else {'
+				decompilerShadow[+x[+op[1] - 2][1] - 1] = '}#'
+			} else {
+				decompilerShadow[opidx] = 'if $pop(0) {'
+				decompilerShadow[+op[1]] = '}'
+			}
+		}
+	}
+
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'call' || op[0] == 'call.meth') {
+			// call
+			// $push($pop(2)($pop(0), $pop(1)))
+			decompilerShadow[opidx] = `$push($pop(${op[1]})(${' '
+				.repeat(+op[1])
+				.split('')
+				.map((_, i) => `$pop(${i})`)
+				.reverse()
+				.join(', ')}))`
+		}
+	}
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'tuple') {
+			// call
+			// $push($pop(2)($pop(0), $pop(1)))
+			decompilerShadow[opidx] = `$push((${' '
+				.repeat(+op[1])
+				.split('')
+				.map((_, i) => `$pop(${i})`)
+				.reverse()
+				.join(', ')},))`
+		}
+	}
+	for (let opidx in x) {
+		let op = x[opidx]
+		if (op[0] == 'list') {
+			// call
+			// $push($pop(2)($pop(0), $pop(1)))
+			decompilerShadow[opidx] = `$push([${' '
+				.repeat(+op[1])
+				.split('')
+				.map((_, i) => `$pop(${i})`)
+				.reverse()
+				.join(', ')}])`
+		}
+	}
+
+	let stack = []
+	let i = -1
+	let tablevel = tlba
+	for (let _e of decompilerShadow) {
+		i++
+		let o = evalopwithstack(decompilerShadow, i, x, stack, tablevel)
+		stack = o.stack
+		tablevel = o.tablevel
+	}
+}
+
+let totalinput = readFileSync(process.argv[2]).toString()
+let cur: string = '_'
+for (let l of totalinput.split('\n')) {
+	if (l.trim() == '') continue
+	if (!l.startsWith('.') && l.endsWith(':')) {
+		// we are an fn
+		grps.set(l.slice(0, -1), '')
+		cur = l.slice(0, -1)
+		ids.push(cur)
+	} else {
+		grps.set(cur, grps.get(cur) + '\n' + l)
+	}
+}
+decompile_text(grps.get('<module>'), 0)
